fix priv activation

+++ vowpalwabbit/global_data.h
@@ -166,7 +166,7 @@ public:
   bool privacy_activation = false;
   // this is coupled with the bitset size in array_parameters which needs to be determined at compile time
   size_t feature_bitset_size = 32;
-  size_t privacy_activation_threshold = 10;
+  uint32_t privacy_activation_threshold = 10;
 #endif
 
 #ifdef BUILD_FLATBUFFERS
